♻️  fix drawer menu using redux
